+++ applications/power/power_service/power.h
