+++ applications/power/power_service/power_api.c
