+++ applications/power/power_service/views/power_off.h
